Fetch tile entities from chunks instead of world
When looking up tile entities for a chunk to send to a player we currently
loop through every tile entity in the world checking if it is within the
bounds of the relevant chunk. Instead of doing this we can just use the
tile entities list stored in the chunk to avoid this costly searching

Thanks to amaranth (CraftBukkit) and mikeprimm (Spigot) for the original
implementations. Based on:

Bukkit/CraftBukkit@e51c31cea73f3f48b8311657ff55b29fc4f3e451